+++ HISTORY.md
@@ -29,7 +29,7 @@
     * new "get pen attribute" reporter
     * new "write" command in pen category (used to be "label" in tools)
     * new "numbers", "is empty", "map","keep", "find", "combine" and "for each" primitives in list category
-    * new JIT-compiler "blitz-HOF" primitives for "map", "keep" & "combine" via "compile"
+    * new JIT-compiler "blitz-HOF" primitives for "map", "keep", "find" & "combine" via "compile"
     * new "for" loop and "if then else" reporter primitives in the Control category
     * added "neg", "lg" (log2) and "2^" selectors to monadic function reporter in Operators
     * added "^" reporter (power of) in the Operators category
@@ -82,6 +82,7 @@
 
 ### 2019-05-29
 * Threads, Objects: added "Find First" primitive to lists category
+* Blocks, Thread, Objects: added "blitz" version of FIND
 
 ### 2019-05-28
 * Maps: added various different tile hosts